Extracted default stack builder
